third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/box-shadow-offset-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['box-shadow-offset'] = "0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "0 0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "1px 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "1em 2em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "-1px -2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "calc(1em + 2px) calc(3rem + 4vw)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['box-shadow-offset'] = "1px 2px, 3px 4px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
